refactoring the mongoDB models
