+++ routes/genres.js
@@ -1,19 +1,6 @@
+const { Genre, validate } = require("../models/genre");
 const express = require("express");
 const router = express.Router();
-const mongoose = require("mongoose");
-const Joi = require("joi");
-
-const Genre = mongoose.model(
-  "Genre",
-  new mongoose.Schema({
-    name: {
-      type: String,
-      required: true,
-      minlength: 5,
-      maxlength: 50,
-    },
-  })
-);
 
 router.get("/", async (req, res) => {
   const genres = await Genre.find().sort("name");
@@ -27,7 +14,7 @@ router.get("/:id", async (req, res) => {
 });
 
 router.post("/", async (req, res) => {
-  const { error } = validateGenre(req.body);
+  const { error } = validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
   let genre = new Genre({ name: req.body.name });
@@ -37,7 +24,7 @@ router.post("/", async (req, res) => {
 });
 
 router.put("/:id", async (req, res) => {
-  const { error } = validateGenre(req.body);
+  const { error } = validate(req.body);
   if (error) return res.status(400).send(error.details[0].message);
 
   const genre = await Genre.findByIdAndUpdate(
@@ -62,12 +49,5 @@ router.delete("/:id", async (req, res) => {
 
   res.send(genre);
 });
-
-const validateGenre = (genre) => {
-  const schema = {
-    name: Joi.string().required(),
-  };
-  return Joi.validate(genre, schema);
-};
 
 module.exports = router;
